requirements.txt
Django~=5.1.4
django-allauth~=65.3.1
openpyxl~=3.1.2
emoji~=2.14.1
python-decouple~=3.8
django-phonenumber-field~=8.0.0
django-modeltranslation~=0.19.11